+++ src/index.ts
@@ -1,6 +1,6 @@
 "use strict";
 
-import { get, post } from "./lib/api-request";
+import { get, post, setUserAgent } from "./lib/api-request";
 import {
   setTaskDelayTime,
   resetTaskDelayTime,
@@ -76,7 +76,7 @@ function getSymbolUrl(symbol: string): string {
 
   if (!url) {
     throw new Error(
-      `Symbol "${symbol}" not found. The scryfall-client module may need an update.`,
+      `Symbol "${symbol}" not found. The scryfall-client module may need an update.`
     );
   }
 
@@ -92,6 +92,7 @@ function wrap(body: ApiResponse): Model {
 }
 
 export = {
+  setUserAgent,
   setApiRequestDelayTime,
   resetApiRequestDelayTime,
   setTextTransform,