+++ src/lib/api-request/index.ts
@@ -1,9 +1,11 @@
 import sendApiRequest from "../../lib/api-request/send-request-to-api";
 import type { AnyJson } from "../../types/json";
 
+export { setUserAgent } from "./user-agent";
+
 export function get<T>(
   url: string,
-  query?: Record<string, string>,
+  query?: Record<string, string>
 ): Promise<T> {
   return sendApiRequest({
     endpoint: url,
@@ -14,7 +16,7 @@ export function get<T>(
 
 export function post<T>(
   url: string,
-  body?: Record<string, AnyJson>,
+  body?: Record<string, AnyJson>
 ): Promise<T> {
   return sendApiRequest({
     endpoint: url,