+++ src/lib/api-request/user-agent.ts
@@ -1,0 +1,13 @@
+let userAgent = "";
+
+export function getUserAgent() {
+  return userAgent;
+}
+
+export function setUserAgent(agent: string) {
+  userAgent = agent;
+}
+
+export function resetUserAgent() {
+  setUserAgent("");
+}