+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 _Potentially Breaking Changes_
 
 - Drop support for Node < 22
+- Require using `setUserAgent` when running package on a node server
 
 # 0.23.1